+++ src/common/gitProvider/gitProviderRoot.ts
@@ -11,6 +11,11 @@ export abstract class GitProviderRoot {
     throw new SfdxError("getLabel should be implemented on this call");
   }
 
+  public async getBranchDeploymentCheckId(gitBranch: string): Promise<string> {
+    uxLog(this, `Method getBranchDeploymentCheckId(${gitBranch}) is not implemented yet on ${this.getLabel()}`);
+    return null;
+  }
+
   public async postPullRequestMessage(prMessage: PullRequestMessageRequest): Promise<PullRequestMessageResult> {
     uxLog(this, c.yellow("Method postPullRequestMessage is not yet implemented on " + this.getLabel() + " to post " + JSON.stringify(prMessage)));
     return { posted: false, providerResult: { error: "Not implemented in sfdx-hardis" } };
@@ -26,24 +31,5 @@ export abstract class GitProviderRoot {
       prResult = { posted: false, providerResult: { error: e } };
     }
     return prResult;
-  }
-
-  protected async getParentMergeRequestId(): Promise<number> {
-    return null;
-  }
-
-  protected async getPipelineId(): Promise<string> {
-    uxLog(this, c.yellow("Method getPipelineId is not yet implemented on " + this.getLabel()));
-    return null;
-  }
-
-  protected async getJobsFromPipeline(): Promise<string> {
-    uxLog(this, c.yellow("Method getJobsFromPipeline is not yet implemented on " + this.getLabel()));
-    return null;
-  }
-
-  protected async getDeployId(): Promise<string> {
-    uxLog(this, c.yellow("Method getDeployId is not yet implemented on " + this.getLabel()));
-    return null;
   }
 }