+++ src/common/utils/deployUtils.ts
@@ -71,7 +71,7 @@ export async function forceSourcePush(scratchOrgAlias: string, commandThis: any,
     uxLog(this, c.red("\n" + errLog));
     uxLog(
       commandThis,
-      c.yellow(c.bold(`You may${tips.length > 0 ? " also" : ""} copy-paste errors on google to find how to solve the push issues :)`))
+      c.yellow(c.bold(`You may${tips.length > 0 ? " also" : ""} copy-paste errors on google to find how to solve the push issues :)`)),
     );
     elapseEnd("force:source:push");
     throw new SfdxError("Deployment failure. Check messages above");
@@ -154,7 +154,7 @@ export async function forceSourceDeploy(
   testlevel = "RunLocalTests",
   debugMode = false,
   commandThis: any = this,
-  options: any = {}
+  options: any = {},
 ): Promise<any> {
   elapseStart("all deployments");
   const splitDeployments = await buildDeploymentPackageXmls(packageXmlFile, check, debugMode, options);
@@ -172,9 +172,9 @@ export async function forceSourceDeploy(
         commandThis,
         c.cyan(
           `Skipped ${c.bold(deployment.label)} deployment because package.xml is empty or contains only standalone parent items.\n${c.grey(
-            c.italic("This may be related to filtering using packageDeployOnce.xml or packageDeployOnChange.xml")
-          )}`
-        )
+            c.italic("This may be related to filtering using packageDeployOnce.xml or packageDeployOnChange.xml"),
+          )}`,
+        ),
       );
       elapseEnd(`deploy ${deployment.label}`);
       continue;
@@ -189,8 +189,39 @@ export async function forceSourceDeploy(
     if (deployment.packageXmlFile) {
       uxLog(
         commandThis,
-        c.cyan(`${check ? "Simulating deployment of" : "Deploying"} ${c.bold(deployment.label)} package: ${deployment.packageXmlFile} ...`)
+        c.cyan(`${check ? "Simulating deployment of" : "Deploying"} ${c.bold(deployment.label)} package: ${deployment.packageXmlFile} ...`),
       );
+      // Try QuickDeploy
+      if (check === false && (process.env?.SFDX_HARDIS_QUICK_DEPLOY || "") !== "false") {
+        const deploymentCheckId = await GitProvider.getDeploymentCheckId();
+        if (deploymentCheckId) {
+          const quickDeployCommand =
+            `sfdx force:source:deploy` +
+            ` --validateddeployrequestid ${deploymentCheckId} ` +
+            (options.targetUsername ? ` --targetusername ${options.targetUsername}` : "") +
+            ` --wait ${process.env.SFDX_DEPLOY_WAIT_MINUTES || "60"}` +
+            ` --verbose` +
+            (process.env.SFDX_DEPLOY_DEV_DEBUG ? " --dev-debug" : "");
+          const quickDeployRes = await execSfdxJson(quickDeployCommand, commandThis, {
+            output: true,
+            debug: debugMode,
+            fail: false,
+          });
+          if (quickDeployRes.status === 0) {
+            uxLog(commandThis, c.green(`Successfully processed QuickDeploy for deploymentId ${deploymentCheckId}`));
+            uxLog(commandThis, c.yellow("If you do not want to use QuickDeploy feature, define env variable SFDX_HARDIS_QUICK_DEPLOY=false"));
+            break;
+          } else {
+            uxLog(
+              commandThis,
+              c.yellow(
+                `Unable to perform QuickDeploy for deploymentId ${deploymentCheckId}.\n${quickDeployRes.errorMessage}\n Switching back to effective deployment.`,
+              ),
+            );
+          }
+        }
+      }
+      // No QuickDeploy Available, or QuickDeploy failing : try full deploy
       const deployCommand =
         `sfdx force:source:deploy -x "${deployment.packageXmlFile}"` +
         ` --wait ${process.env.SFDX_DEPLOY_WAIT_MINUTES || "60"}` +
@@ -214,10 +245,14 @@ export async function forceSourceDeploy(
       } catch (e) {
         const { tips, errLog } = await analyzeDeployErrorLogs(e.stdout + e.stderr, true, { check: check });
         uxLog(commandThis, c.red(c.bold("Sadly there has been Deployment error(s)")));
-        uxLog(this, c.red("\n" + errLog));
+        if (process.env?.SFDX_HARDIS_DEPLOY_ERR_COLORS === "false") {
+          uxLog(this, "\n" + errLog);
+        } else {
+          uxLog(this, c.red("\n" + errLog));
+        }
         uxLog(
           commandThis,
-          c.yellow(c.bold(`You may${tips.length > 0 ? " also" : ""} copy-paste errors on google to find how to solve the deployment issues :)`))
+          c.yellow(c.bold(`You may${tips.length > 0 ? " also" : ""} copy-paste errors on google to find how to solve the deployment issues :)`)),
         );
         await displayDeploymentLink(e.stdout + e.stderr, options);
         elapseEnd(`deploy ${deployment.label}`);
@@ -305,7 +340,7 @@ async function displayDeploymentLink(rawLog: string, options: any) {
     {
       fail: true,
       output: false,
-    }
+    },
   );
   uxLog(this, c.yellowBright(`Open deployment status page in org with url: ${c.bold(c.greenBright(openRes?.result?.url))}`));
 }
@@ -439,7 +474,7 @@ export async function buildDeployOnChangePackageXml(debugMode: boolean, options:
       fail: true,
       output: true,
       debug: debugMode,
-    }
+    },
   );
 
   // Do not call delta if no updated file has been retrieved
@@ -482,7 +517,7 @@ export async function removePackageXmlContent(
   packageXmlFile: string,
   packageXmlFileToRemove: string,
   removedOnly = false,
-  options = { debugMode: false, keepEmptyTypes: false }
+  options = { debugMode: false, keepEmptyTypes: false },
 ) {
   if (removedOnly === false) {
     uxLog(this, c.cyan(`Removing ${c.green(path.basename(packageXmlFileToRemove))} content from ${c.green(path.basename(packageXmlFile))}...`));
@@ -490,8 +525,8 @@ export async function removePackageXmlContent(
     uxLog(
       this,
       c.cyan(
-        `Keeping ${c.green(path.basename(packageXmlFileToRemove))} content from ${c.green(path.basename(packageXmlFile))} (and remove the rest)...`
-      )
+        `Keeping ${c.green(path.basename(packageXmlFileToRemove))} content from ${c.green(path.basename(packageXmlFile))} (and remove the rest)...`,
+      ),
     );
   }
   await removePackageXmlFilesContent(packageXmlFile, packageXmlFileToRemove, {
@@ -515,7 +550,7 @@ export async function deployDestructiveChanges(packageDeletedXmlFile: string, op
     <Package xmlns="http://soap.sforce.com/2006/04/metadata">
         <version>${CONSTANTS.API_VERSION}</version>
     </Package>`,
-    "utf8"
+    "utf8",
   );
   await fs.copy(packageDeletedXmlFile, path.join(tmpDir, "destructiveChanges.xml"));
   const deployDelete =
@@ -542,9 +577,9 @@ export async function deployDestructiveChanges(packageDeletedXmlFile: string, op
       this,
       c.yellow(
         c.bold(
-          "That could be a false positive, as in real deployment, the package.xml deployment will be committed before the use of destructiveChanges.xml"
-        )
-      )
+          "That could be a false positive, as in real deployment, the package.xml deployment will be committed before the use of destructiveChanges.xml",
+        ),
+      ),
     );
     throw new SfdxError("Error while deploying destructive changes");
   }
@@ -568,7 +603,7 @@ export async function deployMetadatas(
     soap: false,
     targetUsername: null,
     tryOnce: false,
-  }
+  },
 ) {
   // Perform deployment
   const deployCommand =
@@ -641,7 +676,7 @@ async function replaceQuickActionsWithDummy() {
     <optionsCreateFeedItem>false</optionsCreateFeedItem>
     <type>LightningComponent</type>
     <width>100</width>
-</QuickAction>`
+</QuickAction>`,
       );
       uxLog(this, c.grey("Backuped and replaced " + quickActionFile));
     }
@@ -694,7 +729,7 @@ export async function buildOrgManifest(targetOrgUsernameAlias, packageXmlOutputF
         fail: true,
         debug: process.env.DEBUG,
         output: true,
-      }
+      },
     );
   } else {
     const tmpDirSfdxProject = await createTempDir();
@@ -712,13 +747,13 @@ export async function buildOrgManifest(targetOrgUsernameAlias, packageXmlOutputF
         cwd: path.join(tmpDirSfdxProject, "sfdx-hardis-blank-project"),
         debug: process.env.DEBUG,
         output: true,
-      }
+      },
     );
   }
   const packageXmlFull = packageXmlOutputFile;
   if (!fs.existsSync(packageXmlFull)) {
     throw new SfdxError(
-      c.red("[sfdx-hardis] Unable to generate package.xml. This is probably an auth issue or a Salesforce technical issue, please try again later")
+      c.red("[sfdx-hardis] Unable to generate package.xml. This is probably an auth issue or a Salesforce technical issue, please try again later"),
     );
   }
   // Add Elements that are not returned by sfdx command
@@ -806,8 +841,8 @@ export async function extractOrgCoverageFromLog(stdout) {
   uxLog(
     this,
     c.italic(
-      c.grey("Unable to get org coverage from results. Maybe try to add --coverageformatters json-summary to your call to force:source:deploy ?")
-    )
+      c.grey("Unable to get org coverage from results. Maybe try to add --coverageformatters json-summary to your call to force:source:deploy ?"),
+    ),
   );
   return null;
 }
@@ -840,7 +875,7 @@ async function checkDeploymentErrors(e, options, commandThis = null) {
   uxLog(this, c.red("\n" + errLog));
   uxLog(
     commandThis,
-    c.yellow(c.bold(`You may${tips.length > 0 ? " also" : ""} copy-paste errors on google to find how to solve the metadata deployment issues :)`))
+    c.yellow(c.bold(`You may${tips.length > 0 ? " also" : ""} copy-paste errors on google to find how to solve the metadata deployment issues :)`)),
   );
   await displayDeploymentLink(e.stdout + e.stderr, options);
   // Post pull requests comments if necessary